添加导航功能以及bootstrap样式

+++ assets/js/app.js
@@ -1,58 +1,109 @@
 'use strict';
 
-$(function() {
+jQuery(function($) {
 	var tagsSplitter = /\s*,\s*/;
 	var user = null;
-	
+	var $modal;
+
+	var modal = {
+		$modal: null,
+
+		init: function() {
+			if (!this.$modal) {
+				$('body').append($('#modal-template').html());
+				$modal = $('#modal');
+			}
+		},
+
+		setContent: function(el) {
+			$('.modal-body', this.$modal).append($(el));
+		},
+
+		setTitle: function(title) {
+			$('model-header h3', this.$modal).html(title);
+		}
+	};
+
 	var SignView = Backbone.View.extend({
 		el: '#sign',
-		
+
+		events: {
+			'click #login': 'login',
+			'click #logout': 'logout',
+		},
+
 		render: function() {
 			var html;
 			if (!user)
-				html ='<a href="#/login">登录</a>';
+				html ='<p class="navbar-text"><a href="#" id="login">登录</a></p>';
 			else
-				html = '欢迎你' + _.escape(user.name) +'！ <a href="#/logout">退出</a>';
-			
+				html ='<p class="navbar-text">欢迎你' + _.escape(user.name) +'！ <a href="#" id="logout">退出</a></p>';
+
 			this.$el.html(html);
-		}
+			return this;
+		},
+
+		login: function(e) {
+			this.$el.hide();
+			appView.loginView.render().$el.show();
+			e.preventDefault();
+		},
+
+		logout: function(e) {
+			var self = this;
+			$.ajax({
+				url: 'ajax.php?r=logout',
+				success: function(resp) {
+					if (resp == 1) {
+						user = null;
+						self.render();
+						appView.formView.$el.hide();
+						appView.trigger('afterLogout');
+					}
+				}
+			});
+			e.preventDefault();
+		},
 	});
-	
+
 	var Category = Backbone.Model.extend({
 		defaults: {
 			name: '',
 			count: 0
 		}
 	});
-	
+
 	var Categories = Backbone.Collection.extend({
 		model: Category,
-		
+
 		url: 'ajax.php?r=categories',
-		
+
 		parse: function(resp, options) {
 			return resp.models;
 		}
 	});
-	
+
 	var CategoriesView = Backbone.View.extend({
 		el: '#categories',
-		
+
 		initialize: function() {
 			this.listenTo(categories, 'reset', this.render);
 		},
-		
+
 		render: function(categories) {
-			var html = this.$el.html();
+			var total = 0;
+			var html = '';
 			categories.each(function(category) {
-				html += '<li><a href="#/category/' + category.escape('id') + '">' + category.escape('name') + '</a>' + '<span>' + category.escape('count') + '</span></li>';
+				html += '<li><a class="category-' + category.escape('id') + '"  href="#/category/' + category.escape('id') + '">' + category.escape('name') + '<span>' + category.escape('count') + '</span></a></li>';
+				total += window.parseInt(category.get('count'));
 			});
+			html = '<li><a href="#/">所有<span>' + total + '</span></a>' + html;
 			this.$el.html(html);
 		}
 	});
-	
+
 	var categories = new Categories();
-	
+
 	var Bookmark = Backbone.Model.extend({
 		defaults: {
 			title: '',
@@ -60,9 +111,9 @@ $(function() {
 			category: {id: 0, name: ''},
 			tags: []
 		},
-		
+
 		urlRoot: 'ajax.php?r=post',
-		
+
 		parse: function(resp, options) {
 			if (resp.error == '0') {
 				return resp.model;
@@ -70,73 +121,89 @@ $(function() {
 				return resp;
 			}
 		},
-		
+
 		validate: function(attrs, options) {
 			if (!attrs.url.trim())
 				return 'Url不能为空';
-			
+
 			if (!/^https?:\/\/(([A-Z0-9][A-Z0-9_-]*)(\.[A-Z0-9][A-Z0-9_-]*)+)/i.test(attrs.url))
 				return 'Url格式不正确';
 		},
-		
+
 		initialize: function() {
 			this.on('invalid', function(model, error){
 				alert(error);
 			});
-			
+
 			this.on('error', function(model, xhr) {
 				if (xhr.status == 404)
 					alert('没有权限访问，请登录');
 			});
 		}
 	});
-	
+
 	var Bookmarks = Backbone.Collection.extend({
+		query: {},
+
 		total: 0,
-		
+
 		page: 1,
-		
+
+		maxPage: 1,
+
+		sectionNum: 5,
+
 		model: Bookmark,
-		
+
 		url: 'ajax.php?r=posts',
-		
+
 		//localStorage: new Backbone.LocalStorage('bookmarks-backbone'),
-		
+
 		parse: function(resp, options) {
 			this.total = resp.total;
+			this.maxPage = Math.max(Math.ceil(this.total/20), 1);
 			return resp.models;
 		}
 	});
-	
+
 	var bookmarks = new Bookmarks();
-	
+
 	var BookmarkView = Backbone.View.extend({
 		tagName: 'div',
-		
+
 		attributes: {
 			'class': 'item',
 		},
-		
+
 		template: _.template($('#item-template').html()),
-		
+
 		events: {
 			'click .edit': 'edit',
 			'click .destory': 'clear',
 		},
-		
+
 		initialize: function() {
 			this.listenTo(this.model, 'change', this.render);
 			this.listenTo(this.model, 'destroy', this.remove);
+			this.listenTo(this.model, 'change:category', this.changeCategory);
 		},
-		
+
 		render: function() {
 			this.$el.html(this.template(this.model.toJSON()));
-			this.$edit = this.$('.edit').hide();
-			this.$destory = this.$('.destory').hide();
+			this.$edit = this.$('.edit');
+			this.$destory = this.$('.destory');
+			if (!user) {
+				this.$edit.hide();
+				this.$destory.hide();
+			}
 			return this;
 		},
-		
-		edit: function() {
+
+		changeCategory: function(model) {
+			categories.fetch({'reset': true});
+		},
+
+		edit: function(e) {
 			var formView = appView.formView;
 			formView.model = this.model;
 			formView.listenTo(this.model, 'destroy', formView.reset);
@@ -147,42 +214,48 @@ $(function() {
 			formView.$url.val(this.model.get('url'));
 			formView.$category.val(this.model.get('category').id);
 			formView.$tags.val(this.model.get('tags').join(','));
+			//console.log($modal);
+			//$('.modal-body', $modal).append(formView.$el);
+			//$modal.modal();
 			window.location.href = '#post';
+			e.preventDefault();
 		},
-		
-		clear: function () {
-			this.model.destroy({wait: true});
+
+		clear: function(e) {
+			if (window.confirm('确定要删除这个书签吗？'))
+				this.model.destroy({wait: true});
+			e.preventDefault();
 		},
-		
+
 		afterLogin: function() {
 			this.$edit.show();
 			this.$destory.show();
 		},
-		
+
 		afterLogout: function() {
 			this.$edit.hide();
 			this.$destory.hide();
 		}
 	});
-	
+
 	var FormView = Backbone.View.extend({
 		tagName: 'form',
-		
+
 		isNewRecord: true,
-		
+
 		template: _.template($('#form-template').html()),
-		
+
 		events: {
 			'click #submit': 'submit',
 			'click #cannel': 'reset',
 		},
-		
+
 		initialize: function() {
 			this.listenTo(categories, 'reset', this.renderCategories);
 			this.listenTo(bookmarks, 'add', this.reset);
 			this.listenTo(bookmarks, 'reset', this.reset);
 		},
-		
+
 		render: function() {
 			this.$el.html(this.template());
 			this.$title = this.$('#title');
@@ -194,7 +267,7 @@ $(function() {
 			$('#post').append(this.$el);
 			return this;
 		},
-		
+
 		renderCategories: function(categories) {
 			var html='';
 			categories.each(function(category){
@@ -202,7 +275,7 @@ $(function() {
 			});
 			this.$category.html(html);
 		},
-		
+
 		reset: function() {
 			delete this.model;
 			this.$title.val('');
@@ -212,7 +285,7 @@ $(function() {
 			this.$submit.html('提交');
 			this.$cannel.hide();
 		},
-		
+
 		submit: function(e) {
 			if (!this.model) {
 				bookmarks.create(this.newAttributes(), {wait: true});
@@ -224,12 +297,12 @@ $(function() {
 			}
 			e.preventDefault();
 		},
-		
+
 		newAttributes: function() {
 			var url = this.$url.val().trim();
 			var title = this.$title.val().trim();
 			var category = {id: this.$category.val(), name: $('option:selected', this.$category).html()};
-			
+
 			return {
 				title: title ? title : url,
 				url: url,
@@ -238,22 +311,25 @@ $(function() {
 			};
 		}
 	});
-	
+
 	var LoginView = Backbone.View.extend({
 		tagName: 'form',
-		
+
+		className: 'navbar-form pull-right',
+
 		events: {
-			'click button': 'login',
+			'click button': 'submit',
 		},
-		
+
 		render: function() {
 			this.$el.html($('#login-template').html());
-			$('#sign').after(this.$el);
 			this.$username = this.$('#username');
 			this.$password = this.$('#password');
+			$('#sign').after(this.$el);
+			return this;
 		},
-		
-		login: function(e) {
+
+		submit: function(e) {
 			var self = this;
 			$.ajax({
 				url: 'ajax.php?r=login',
@@ -262,7 +338,7 @@ $(function() {
 				success: function(resp) {
 					if (resp.error == 0) {
 						user = resp.user;
-						appView.signView.render();
+						appView.signView.render().$el.show();
 						self.$el.hide();
 						appView.formView.$el.show();
 						appView.trigger('afterLogin');
@@ -274,34 +350,78 @@ $(function() {
 			e.preventDefault();
 		}
 	});
-	
+
 	var AppView = Backbone.View.extend({
+
 		isFetch: false,
-		
+
 		el: '#bookmarkapp',
-		
+
 		formTemplate: _.template($('#form-template').html()),
 
 		initialize: function() {
+			var self = this;
+			var pagination = $('#pagination');
+
+			var loading = $('.loading', this.$el);
 			this.listenTo(bookmarks, 'add', this.addOne);
-			this.listenTo(bookmarks, 'reset', this.pagination);
 			this.listenTo(bookmarks, 'reset', this.addAll);
+			this.listenTo(bookmarks, 'reset', function() {
+				if (bookmarks.query['category_id']) {
+					category = categories.findWhere({'id': bookmarks.query['category_id']});
+					if (category) {
+						$('h2', self.$el).html('分类：' + category.escape('name'));
+					}
+				} else if (bookmarks.query['tag']) {
+					$('h2', self.$el).html('标签：' + _.escape(bookmarks.query['tag']));
+				} else
+					$('h2', self.$el).html('');
+			});
+			this.listenTo(bookmarks, 'request', function() {
+				loading.show();
+				pagination.html('');
+			});
+			this.listenTo(bookmarks, 'sync', function() {
+				loading.hide();
+				self.pagination();
+			});
 			this.listenTo(categories, 'reset', this.addCategories);
+			this.listenTo(bookmarkRouter, 'route', function(route, params){
+				var tag, category;
+				if (_.indexOf(['show', 'showByCategory', 'showByTag'], route) !== -1) {
+					self.isFetch = true;
+				}
+			});
+
 			this.formView = new FormView();
 			this.signView = new SignView();
 			this.loginView = new LoginView();
 			this.categoiresView = new CategoriesView();
-			
-			//if (!this.isFetch)
-				//bookmarks.fetch({reset: true, data:{'page':1}});
-			
-			$(window).scroll(function(){
+
+			var throttled = _.throttle(function(){
+				if (bookmarks.maxPage <= bookmarks.page || (bookmarks.page % bookmarks.sectionNum === 0))
+					return;
+
 				var scrollHeight = $(window).scrollTop() + $(window).height();
-			});
-			
+
+				if (scrollHeight >= pagination.offset()['top']) {
+					var data = {};
+					data['page'] = ++bookmarks.page;
+
+					if (bookmarks.query['category_id'])
+						data['category_id'] = bookmarks.query['category_id'];
+					else if (bookmarks.query['tag'])
+						data['tag'] = bookmarks.query['tag'];
+
+					bookmarks.fetch({data: data});
+				}
+			}, 100);
+
+			$(window).scroll(throttled);
+
 			this.checkLogin();
 		},
-		
+
 		checkLogin: function()
 		{
 			var self = this;
@@ -318,13 +438,15 @@ $(function() {
 				}
 			});
 		},
-		
+
 		render: function() {
 			this.signView.render();
 			this.formView.render().$el.hide();
-			categories.fetch({reset:true});
+			categories.fetch({'reset': true});
+			if (!this.isFetch)
+				bookmarks.fetch({reset: true, data:{'page':1}});
 		},
-		
+
 		addOne: function(bookmark) {
 			if (!bookmark.isNew()) {
 				var view = new BookmarkView({model: bookmark});
@@ -334,107 +456,115 @@ $(function() {
 				$('#items').append(view.render().el);
 			}
 		},
-		
+
 		addAll: function(bookmarks) {
 			this.$('#items').html('');
 			bookmarks.each(this.addOne, this);
 			this.isFetch = true;
 		},
-		
-		pagination: function(bookmarks) {
-			var page = bookmarks.page;
-			var total = bookmarks.total;
-			var maxPage = Math.ceil(total/20);
-			var pagination = [];
+
+		pagination: function() {
+			var sectionNum = bookmarks.sectionNum;
+			var page = Math.ceil(bookmarks.page / sectionNum);
+			var maxPage = Math.ceil(bookmarks.maxPage / sectionNum);
+
+			if (bookmarks.page %2 !== 0) {
+				if (maxPage !== page || page == 1)
+					return;
+			}
+
+
+			var pagination = {};
+			var url;
+			var html = '';
+
 			if (maxPage > 1) {
 				if (page == 1) {
-					pagination.push({'下一页': page+1});
-				} else if (page > 1 && page < maxPage-1) {
-					pagination.push({'上一页': page-1});
-					pagination.push({'下一页': page+1});
+					pagination = {'下一页': page*sectionNum + 1};
+				} else if (page > 1 && page < maxPage) {
+					pagination={
+						'上一页': bookmarks.page - page*sectionNum + 1,
+						'下一页': bookmarks.page + 1
+					};
 				} else {
-					pagination.push({'上一页': page-1});
+					pagination = {'上一页': (page-2) * sectionNum + 1};
 				}
+
+				if (bookmarks.query['category_id'])
+					url = '#/category/' + _.escape(bookmarks.query['category_id']) + '/page/:page';
+				else if (bookmarks.query['tag'])
+					url = '#/tag/' + _.escape(bookmars.query['tag']) + '/page/:page';
+				else
+					url = '#/page/:page';
+
+				_.each(pagination, function(value, key){
+					html += '<a href="' + url.replace(':page', value) + '">' + key + '</a>';
+				});
+
+				$('#pagination').html(html);
 			}
 		}
 	});
-	
+
 	var BookmarkRouter = Backbone.Router.extend({
 		routes: {
 			'': 'show',
-			'logout': 'logout',
-			'login': 'login',
 			'page/:page': 'show',
 			'category/:id': 'showByCategory',
 			'category/:id/page/:page': 'showByCategory',
 			'tag/:tag': 'showByTag',
 			'tag/:tag/page/:page': 'showByTag'
 		},
-		
-		logout: function() {
-			$.ajax({
-				url: 'ajax.php?r=logout',
-				success: function(resp) {
-					if (resp == 1) {
-						user = null;
-						appView.signView.render();
-						appView.formView.$el.hide();
-						appView.trigger('afterLogout');
-					}
-				}
-			});
-		},
-		
-		login: function() {
-			appView.loginView.render();
-			appView.loginView.$el.show();
-		},
-		
+
 		show: function(page) {
 			var data = {};
 			page = parseInt(page);
 			if (!page)
 				page = 1;
 			bookmarks.page = page;
+			delete bookmarks.query['tag'];
+			delete bookmarks.query['category_id'];
 			bookmarks.fetch({reset: true, data: {page: page}});
 		},
-		
+
 		showByCategory: function (category_id, page) {
 			var data = {};
 			if (category_id)
 				data['category_id'] = category_id;
-			
+
 			page = parseInt(page);
 			if (!page)
 				page = 1;
 			data['page'] = page;
 			bookmarks.page = page;
+			bookmarks.query['category_id'] = category_id;
+			delete bookmarks.query['tag'];
 			bookmarks.fetch({reset: true, data: data});
 		},
-		
+
 		showByTag: function (tag, page) {
 			var data = {};
 			if (tag)
 				data['tag'] = tag;
-			
+
 			page = parseInt(page);
 			if (!page)
 				page = 1;
 			data['page'] = page;
 			bookmarks.page = page;
+			bookmarks.query['tag'] = tag;
+			delete bookmarks.query['category_id'];
 			bookmarks.fetch({reset: true, data: data});
 		}
 	});
-	
+
 	Backbone.emulateHTTP = true;
 	Backbone.emulateJSON = true;
 
-	var appView = new AppView();
 	var bookmarkRouter = new BookmarkRouter();
-	Backbone.history.on('route', function(){
-		console.log(arguments);
-	});
+	var appView = new AppView();
+
 	Backbone.history.start();
-	
+
 	appView.render();
 });
